refactor: move socket setting from app.js to /utils/socket.js
